Added global states for the manage of reviews and searched reviews
