+++ MovieReviewer/src/Header/Navbar/Navbar.jsx
@@ -4,14 +4,14 @@ import SearchBar from '../SearchBar/SearchBar'
 import UserSlot from '../UserSlot/UserSlot'
 import './Navbar.css'
 
-function Navbar() {
+function Navbar({searchValue, onChange, onClick}) {
     return (
         <nav>
             {/* Go to the homepage by clicking on the logo */}
             <Link to='/'>
                 <img src={movieLogo} className="logo" alt="Movie logo" />
             </Link>
-            <SearchBar />
+            <SearchBar {...{searchValue, onChange, onClick}}/>
 
             {/* Go to the user profile by clicking on the user slot */}
             <Link to='/profile/user1' style={{ textDecoration: 'none' }}>
