+++ MovieReviewer/src/Header/SearchBar/SearchBar.jsx
@@ -1,19 +1,14 @@
-import { useState } from 'react'
 import movieSearchBar from '../../assets/logos/movieSearchBar.svg'
 import searchBarIcon from '../../assets/logos/searchBarIcon.svg'
 import userSolid from '../../assets/logos/userSolid.svg'
 import './SearchBar.css'
 
-function SearchBar() {
-    const [searchValue, setSearchValue] = useState("")
-
-    const onChange = (e) => {
-        setSearchValue(e.value)
-    }
-
+function SearchBar({searchValue, onChange, onClick}) {
     return (
         <div className="searchBar">
-            <img src={searchBarIcon} className="logo" alt="Search icon" />
+            <button id="searchButton" onClick={onClick}>
+                <img src={searchBarIcon} className="logo" alt="Search icon" />
+            </button>
             <input type="text" value={searchValue} onChange={onChange}/>
             <img src={movieSearchBar} className="logo borderLeft borderRigth" alt="Movie search bar icon"/>
             <img src={userSolid} className="logo" alt="User solid icon"/>
